renamed repository directory to "repositoryTypes"
