+++ phoenicis-repository/src/test/java/org/phoenicis/repository/repositoryTypes/CachedRepositoryTest.java
@@ -16,13 +16,13 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package org.phoenicis.repository.repository;
+package org.phoenicis.repository.repositoryTypes;
 
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.phoenicis.repository.dto.CategoryDTO;
-import org.phoenicis.repository.repository.CachedRepository;
-import org.phoenicis.repository.repository.Repository;
+import org.phoenicis.repository.repositoryTypes.CachedRepository;
+import org.phoenicis.repository.repositoryTypes.Repository;
 
 import java.util.Arrays;
