+++ phoenicis-repository/src/test/java/org/phoenicis/repository/repositoryTypes/MultipleRepositoryTest.java
@@ -16,12 +16,12 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package org.phoenicis.repository.repository;
+package org.phoenicis.repository.repositoryTypes;
 
 import org.phoenicis.repository.dto.CategoryDTO;
 import org.junit.Test;
-import org.phoenicis.repository.repository.MultipleRepository;
-import org.phoenicis.repository.repository.Repository;
+import org.phoenicis.repository.repositoryTypes.MultipleRepository;
+import org.phoenicis.repository.repositoryTypes.Repository;
 
 import java.util.Collections;
