+++ phoenicis-repository/src/test/java/org/phoenicis/repository/repositoryTypes/TeeRepositoryTest.java
@@ -16,13 +16,13 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package org.phoenicis.repository.repository;
+package org.phoenicis.repository.repositoryTypes;
 
 import org.phoenicis.repository.dto.ApplicationDTO;
 import org.phoenicis.repository.dto.CategoryDTO;
 import org.junit.Test;
-import org.phoenicis.repository.repository.Repository;
-import org.phoenicis.repository.repository.TeeRepository;
+import org.phoenicis.repository.repositoryTypes.Repository;
+import org.phoenicis.repository.repositoryTypes.TeeRepository;
 
 import java.util.Arrays;
 import java.util.Collections;
