+++ phoenicis-tests/src/main/java/org/phoenicis/tests/MockedRepository.java
@@ -18,9 +18,9 @@
 
 package org.phoenicis.tests;
 
-import org.phoenicis.repository.repository.Repository;
+import org.phoenicis.repository.repositoryTypes.Repository;
 import org.phoenicis.entities.OperatingSystem;
-import org.phoenicis.repository.repository.TeeRepository;
+import org.phoenicis.repository.repositoryTypes.TeeRepository;
 import org.phoenicis.repository.dto.ApplicationDTO;
 import org.phoenicis.repository.dto.CategoryDTO;
 import org.phoenicis.repository.dto.ScriptDTO;
